third_party/WebKit/LayoutTests/inspector/styles/override-and-disable.html

<html>
<head>

<style type="text/css" media="screen">
@media (max-device-width: 478px) and (max-device-height: 799px) {
    #main { background: red; }
}

@media (max-device-width: 480px) and (max-device-height: 800px) {
    #main { background: green; }
}

@media (min-device-width: 481px) and (min-device-height: 801px) {
    #main { background: yellow; }
}

body {
    position: absolute;
    top: 0;
    right: 0;
    bottom: 0;
    left: 0;
    margin: 0;
    font-size: 8px;
}

</style>

<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script>

function getSizes()
{
    return JSON.stringify({
        screen: window.screen.width + "x" + window.screen.height,
        inner: window.innerWidth + "x" + window.innerHeight,
        body: document.body.offsetWidth + "x" + document.body.offsetHeight
    });
}

function test()
{
    var overridesEnabled = false;
    var originalScreenSize;

    function saveSizeCallback(jsonResult)
    {
        var value = JSON.parse(jsonResult.value);
        originalScreenSize = value.screen;
        if (!originalScreenSize || originalScreenSize.indexOf("x") === -1)
            InspectorTest.addResult("Invalid original screen size: " + originalScreenSize + ". Expect runtime failures later on.");

        InspectorTest.selectNodeAndWaitForStyles("main", step0);
    }

    InspectorTest.evaluateInPage("getSizes()", saveSizeCallback);


    function step0()
    {
        overrideAndDumpData(480, 800, step1);
    }

    function step1()
    {
        function callback(jsonResult)
        {
            var value = JSON.parse(jsonResult.value);
            if (value.screen !== originalScreenSize)
                InspectorTest.addResult("Screen size not restored, actual: " + value.screen + ", expected: " + originalScreenSize);
            else
                InspectorTest.addResult("Screen size same as original - OK");
            step2();
        }

        InspectorTest.addResult("Disable PageAgent:");
        PageAgent.disable();
        InspectorTest.evaluateInPage("getSizes()", callback);
    }

    function step2()
    {
        function callback(jsonResult)
        {
            var value = JSON.parse(jsonResult.value);
            if (value.screen !== originalScreenSize)
                InspectorTest.addResult("Screen size not restored, actual: " + value.screen + ", expected: " + originalScreenSize);
            else
                InspectorTest.addResult("Screen size same as original - OK");
            InspectorTest.completeTest();
        }

        InspectorTest.addResult("Enable PageAgent:");
        PageAgent.enable();
        InspectorTest.evaluateInPage("getSizes()", callback);
    }

    function applyOverride(width, height, userCallback)
    {
        function callback(reload, error)
        {
            if (error)
                InspectorTest.addResult("Override: " + width + "x" + height + " => ERROR");
            if (reload && !error) {
                InspectorTest.reloadPage(userCallback);
            } else {
                userCallback();
            }
        }
        
        var enabled = width > 0 && height > 0;
        PageAgent.setDeviceMetricsOverride(width, height, 1, true, callback.bind(null, enabled != overridesEnabled));
        overridesEnabled = enabled;
    }

    function overrideAndDumpData(width, height, callback)
    {
        function finalCallback()
        {
            InspectorTest.addResult("Main style:");
            InspectorTest.dumpSelectedElementStyles(true, false, true);
            callback();
        }

        var gotSizes;
        var gotStyles;
        function stylesCallback()
        {
            if (gotSizes)
                return finalCallback();
            gotStyles = true;
        }

        function sizesCallback()
        {
            gotSizes = true;
            if (gotStyles)
                finalCallback();
        }

        function applyCallback()
        {
            getAndDumpSizes(sizesCallback);
        }

        InspectorTest.addResult("Override: " + width + "x" + height);
        InspectorTest.waitForStyles("main", stylesCallback);
        applyOverride(width, height, applyCallback);
    }

    function getAndDumpSizes(callback)
    {
        function getSizesCallback(jsonResult)
        {
            var result = JSON.parse(jsonResult.value);
            InspectorTest.addResult("Screen from page: " + result.screen);
            InspectorTest.addResult("Window from page: " + result.inner);
            InspectorTest.addResult("Body from page: " + result.body);
            if (callback)
                callback();
        }

        InspectorTest.evaluateInPage("getSizes()", getSizesCallback);
    }
}
</script>
</head>

<body onload="runTest()">
<p>
Tests that disabling page agent cancels device metrics override.
</p>

<div id="main"></div>
</body>
</html>
